requirements.txt
Django==5.0
djangorestframework==3.15
drf-yasg==2.2.0
scikit-learn==1.5
pandas==2.2
numpy
joblib
scipy
requests
